Update github.com/creachadair/jrpc2 to v0.26.1.
In preparation for committing to a stable v1 release for jrpc2 package, I have
made a number of updates to the library to simplify the API a bit. These
changes are discussed in creachadair/jrpc2#46 if you would like to offer input.

This commit updates the vendored package version used here to the latest tag
and updates the existing use for the changes.  These changes should not produce
any functional differences.